Fix bazel build: Remove dep on wire_protocol
Right now we ended dup in an unfortunate state where
the IPC layer still depends on libprotobuf but most
other targets use only protozero.
This causes a weird situation where wire_protocol.proto
is pulled twice in bazel files, once as .gen.cc and once
as .pb.cc, because of the indirect dep of the IPC target.
Adding a temporary exclusion, which will be remove once
aosp/1163503 lands.

Test: CI
Change-Id: I2affe153fdb052689b0926af97eeff70d12f78a5