Protozero-gen: don't depend on .pbzero.* from .gen.*
Before this CL the generated .gen.* files were depending
on the corresponding .pbzero.* files. This created some
coupling between the two (e.g., making assumptions on the
namespace pbzero::). Turns out this coupling is not necessary
and just creates more problems.
The only method that depends on this is ::Serialize() and it
can be easily re-written without depending on the pbzero.h
headers.

The code in the generated .gen.cc changes as follows:

Before this CL:
    void DataSourceConfig::Serialize(perfetto::protos::pbzero::DataSourceConfig* msg) const {
      if (_has_field_[1]) { msg->set_name(name_); }
      if (_has_field_[2]) { msg->set_target_buffer(target_buffer_); }
      if (_has_field_[3]) { msg->set_trace_duration_ms(trace_duration_ms_); }
      if (_has_field_[7]) { msg->set_stop_timeout_ms(stop_timeout_ms_); }

After this CL:
    void DataSourceConfig::Serialize(::protozero::Message* msg) const {
      // Field 1: name
      if (_has_field_[1]) {
        msg->AppendString(1, name_);
      }

      // Field 2: target_buffer
      if (_has_field_[2]) {
        msg->AppendVarInt(2, target_buffer_);
      }

Bug: 132880619
Change-Id: I8f82fd2f9bcf1ebf376222e5471874e5c6676c59