TrackEvent: Add support for debug annotations
This patch makes it possible to attach one or two debug annotations to a
trace event, e.g.:

    TRACE_EVENT_BEGIN("cat", "Name", "color", 0xffaaccee, "density", .5f);

The user can also specify a custom debug annotation object that will
get serialized to the DebugAnnotation proto through a callback. This is
intended for implementing the legacy TracedValue system on top of
TrackEvent.

Bug: 132678367
Change-Id: I769c2eb346d3bdfac50158a333481f83962bfb6b