Switch TRACE_EVENT to something that is only one line of code.
One line is important because some scopes are only one line. For example

{
if (condition)
  TRACE_EVENT(...);  // TRACE_EVENT_BEGIN emitted conditionally here.

}  // TRACE_EVENT_END emitted here.

Would previous emit the TRACE_EVENT_BEGIN only if |condition| was true,
but would always emit a TRACE_EVENT_END in whatever scope was enclosing
the single line if statement. After this change both the BEGIN and END
events are emitted in the scope of the single line if statement as would
be expected. Preventing misuse of the API.

Change-Id: Id3aebd3247a9b22fab032ebfce88528defeae77f