trace_processor: fix incorrect DCHECK and loop condition
We were incorrectly assuming that the BitVector has to be the size of
the column but this is not necessarily true if there are prior
constraints which reduce the space using ranges.

Instead of relying on the size of the column as the upper bound, rely
on the size of the bitvector.

Bug: 145352420
Change-Id: I996053aaf20505f4a045aaae6731ab0efd0235ae