Create a reusable thread_state table
We want to be able to query the thread states when aggregating
CPU information over selected tracks. This means that the previous
method of having separate views for each utid that contain the thread
states no longer works. We really need one view that contains all of
the utids.
However b/145350531 prevents this from being fast enough for the UI.
In the mean time we will create a thread state table. This uses
around 80-90% of the trace size in memory which is not ideal.
However we are saving some memory in aosp/1178075 so end users
shouldn't notice a difference.
Once the bug above is fixed this can be changed to be a view or
in the future we can store in in the trace processor as an
optimised table to save roughly 50% of the memory cost.


Change-Id: I36ccd7cbce58f71f34154d9a2efd1ba65fafea9f