fuzzer: Remove bad CHECK on Send
Unfortunately we can't always assume that the Send succeeds during
fuzzing. Within Send we may end up calling sendmsg multiple times.
If the first sendmsg causes the receiver to hang up on us (which it
can do if we set the frame size too large) then the second sendmsg
will fail causing the whole method to return the error code.

Bug: https://bugs.chromium.org/p/oss-fuzz/issues/detail?id=15770
Change-Id: I6b45418d1f14835f8d92e37eeb5179b2283083c5