Use base::FlatSet for pid tracking in ftrace metadata
The old code was appending to a never-ending vector,
speculating that the last pid inserted would show up
over and over again.
That speculation was wrong and in some cases it would
cause the metadata vector to grow to +16MB.
Use FlatSet, reduce overhead, make memory bounded to
the number of distinct pids.
Also add a bitset cache ahead of the FlatSet. It makes
AddPid() even faster, saving 1% to 2% on very large
workloads.

Bug: 145398368
Change-Id: I3baffa772ace8267e49daa6101ed92ff913d67c2