Prevent warnings in amalgamated builds
In the amalgamated build some of the constants generate warnings of the
form:
warning: unused variable 'kInvalidPid' [-Wunused-const-variable]

This is because the code that refers to them is not part of the client
library (e.g. the service binaries or trace processor).
This CL adds a bunch of suppressions and remove actually unused vars.

Bug: 145339279
Change-Id: Id3f393a6fda7e55f43f463cd21bd797be8ab1b54