Allow copying host executable targets in Chromium builds
When the trace processor shell executable is built, it is placed into
the host build subdirectory, which may vary from host to host. To use it
on waterfall, we need it in the same location everywhere, e.g. in the
output directory root.

This CL fixes the issue with copying the file to its final location
on Windows.

Bug: chromium:1028612
Change-Id: I6372ff0ca4622c916bde3dda9ea47c57af6dcd30