trace_processor: remove RowId type from trace_processor
This CL removes the RowId type completely from trace processor. Initially
the type was introduced to uniquely identify a row in trace processor
and to use as a join key. However, with the advent of db tables, we no
longer need this ability so we can get rid of the concept.

As there is still a few dependencies on RowId, this CL also does the
following:
1. Changes ArgsTracker to use TableId + row number to identify which
   row an arg belongs to.
2. Add the concept of a BoundInserter to ArgsTracker. A BoundInserter
   is a class which allows callers to add args to the args tracker without
   specifying the table and row - this is because when the BoundInserter
   is created, it is assigned to a Row. By doing this, we reduce the amount
   of complexity throughout trace processor significantly when dealing with
   args
3. Change SliceTracker to use BoundInserter.
4. Change all callsites throughout trace processor to use BoundInserter
   and ArgsTracker properly.

Change-Id: I5034ebb6e47af1387dec7f4d8919f91e8dd43c97