Add build support for iOS and NaCl for libperfetto.
We want to make it so TRACE_EVENT macros will work on all build
platforms that chromium supports. One way is to make libperfetto build for IOS
and NACL. This patch adds support build OS's for libperfetto and deals with any
current incompatibilities.

Another option we considered is to make a "minlibperfetto" that adds just enough
to ensure that protozero gets generated so includes can not be ifdef'd
in chromium, however this requires a bunch of things to move into minlibperfetto
including perfetto logging and base, which in the end is basically this patch.

Change-Id: Iae04fb7b139bdf4c191862aa8f7d1185015f2869