[Reland] perfetto: merge TaskRunnerThread and base::ThreadTaskRunner
This CL folds the functionality of TaskRunnerThread into
ThreadTaskRunner. Both classes performed essentially the same job.

As ThreadTaskRunner is in base and also has a fix for a subtle
initialization race, we keep that implementation. We just port the
GetThreadCPUTimeNs function which we need for some benchmark code.

We also remove the complex delegate system and replace instead with
wrapper classes of ThreadTaskRunner which each initalize the appropriate
producer/service correctly.

The orignal CL broke CTS build - this patch also changes the jni dep to
be an autogenerated cc file instead of relying on manually updating the
Android.bp.extras

Change-Id: Ia34d7b4a50822692d1c0123455ddd432357507d7