proto/processor: Add process name to ProcessDescriptor
Adds an optional string field, to make it possible for producers to
specify a title for the process that's shown in the UI (if no process
name is provided by system probes).

Also refactors tokenization of the descriptors, so that we can move to
TrackDescriptors while still emitting timestamps using the deprecated
ThreadDescriptor reference timestamp fields (to aid transition in
Chrome).

Bug: 123864183, 130786981
Change-Id: I5f847bb1e02a6b364579a10127af04baab40d542