processor: Support TrackDescriptor's parent_uuid + pid/tid reuse
Refactors tokenization and parsing of TrackDescriptors so that:
 - Only a single ProcessDescriptor track is supported per process.
 - Only a single ThreadDescriptor track is supported per thread.
 - Such a process/thread track is merged with slice-type system events
   into a single track.
 - Other tracks (e.g. async tracks) can be associated with a thread
   or process using the "parent_uuid" field.
 - pid/tid reuse (i.e. new process/thread spawns with same pid/tid
   after old process/thread terminates) is detected based on changing
   uuids of Process/ThreadDescriptor tracks.

To accomplish this, tokenization now only makes reservations for tracks
without actually inserting anything into the track table - since
TrackDescriptors may appear in any order in the file, and we cannot
know the track's type/context until we see its parent's descriptor.

During parsing, the reservations will be resolved and inserted into the
track table. If the uuid of a process's or thread's track changes, we
assume that the pid/tid of the process/thread was reused.

Change-Id: Iec1d019db211773eb15ee517ae0dac1ce87553c9