trace_processor: remove tableid and refactor argstracker to use ids
The last place TableId was being used was in args tracker. Now that we
have proper, strongly typed ids we can use this to disambiguate tables
so do not need to use TableId anymore.

This CL changes ArgsTracker to stop using TableId and instead create
BoundInserters based on the type of the passed in Id. We can thus also
remove the IndexOf calls throughout the codebase and use BoundInserter
for everything with minimal boilerplate allowing us to omit flat_key in
a lot of places.

Context: go/perfetto-tp-refactor
Bug: 135177627
Change-Id: I8fec9e3186fdb87b28ed33d06902a363c8873e60