trace_processor: allow start timestamps to be nullopt for thread/process
This CL allows null start timestamps for threads and processes. This
allows for correct behaviour in the cases where we know that a new thread
was started but we don't know the actual time where it happened - in
these cases we prefer to have noted down that we don't know rather than
making up an approximation.

Change-Id: I9dabef4ea72c691f05768f696d66ad8c1943c571