Enable CreateMayFile to fail gracefuly when faced with errors opening the socket.
Currently if the socket() call fails (with permission denied for example) The
PERFETTO_DCHECK(fd_) will fail since it unconditionally calls the
constructor accepting an fd with the result.

This instead creates the fd and only if it is valid passes it along to
the UnixSocketRaw, otherwise it returns an empty invalid socket.

This has been tested in
https://chromium-review.googlesource.com/c/chromium/src/+/1953718 (patchset 3 vs 4)

Change-Id: I6734fe8b0cfeb3b74522ecb22fab807eca18c92f