trace_processor: allow selecting rowmaps with smaller rowmaps
We were being too strict with the size of the picker RowMap - it doesn't
necessarily have to exactly match the number of rows in this (this case
happens when we have a filter on a sorted column e.g. on timestamp
followed by a filter on another column). Since we will have removed a
bunch of rows from timestamp, we won't necessarily have to create a
BitVector of the size of the column.

Bug: 147867069
Change-Id: I101c09bdfefc047aa1d59542e6efbd20375542d0