Rather then InProcSharedShmemArbiter switch to MaybeSharedMemoryArbiter
We want to expose this for Chrome to be able let the IPC layer to own
both the SharedMemoryArbiter & the SharedMemory. Currently uses the
SharedMemory but needs to create its own SharedMemoryArbiter because it
needs to allow Chrome's data sources to create TraceWriters as needed.
By exposing this in the producer_ipc_client_impl we can share the
already created one.

Change-Id: I7bf72e86fe3d7d1f407f0ed88167226a4f7afdb0