TrackEvent: Reduce trace point binary size overhead
This patch introduces dedicated trace entrypoints for common sets of
trace point arguments. The general pattern is that each non-inlined entrypoint
(TraceForCategory) delegates to a common inlined back-end
(TraceForCategoryImpl), giving the following advantages:

1) Parameter setup for TraceForCategoryImpl happens once in each
   TraceForCategory function instead of at each individual trace point.

2) The actual implementation of the trace point is centralized in
   TraceForCategoryImpl, reducing code duplication and making it easier
   to add other specializations later (e.g., timestamp overrides).

Note that this means it's not possible to fully erase the |arg_function|
for typed trace events, because it crosses a non-inlined boundary (this
was already the case before). The only way to completely unfold
|arg_function| would be to inline the entire DataSource trace point
code, but the main down side of this is allocating ~120 bytes of extra
stack memory (x64) _outside_ the trace point, which isn't desirable.

Bug: 147743479
Change-Id: Ie97d015613ce331e7809e1ccd0ee5561dcf6662b