service: Fix verification of SMB/page sizes
Extracts validation logic for SMB and page sizes out of
SetupDataSource() in preparation of support for producer-provided SMBs.

In the process, simplifies the logic to fall back to default values for
both page & SMB size in case either one of them is invalid.

The old logic had a bug: If you passed a page size that was a multiple
of 4kB, but not a divisible of the default SMB size, the service could
accept it nevertheless. Making the logic more complicated to consider
this case correctly seems fragile, so simplify it instead.

Bug: 147742905
Change-Id: Ibd8572a3e59b1c6c5bc97472e95fa7c74ad359d6