traced_probes: supply shmem size hints when connecting
The primary purpose of traced_probes is ftrace, which is
often relatively high bandwidth. Therefore it's better to
have a bigger shmem, and pack more data per tracing page.

Reminder that the first tracing session pins the shmem config
for the remainder of the traced<->traced_probes connection.
And the config-supplied hints take precedence. So these
values aren't guaranteed.

Bug: 146133021
Change-Id: I85ff6f6477c321a0c062be3b398933e7233867aa