tp: add support for non-Perfetto root relative proto imports in metrics
This CL adds support for metric extension protos to be relative to a
root other than the Perfetto root when importing and extending the
metrics.proto file. This allows e.g. G3 builders to refer to the full
path from the source tree root while our code keeps using the perfetto
relative path.

While we're here also remove the metrics-extra option which is undocumented,
unused and has been subsumed by the path variant of --run-metrics.

Change-Id: I43e377bfc53f3a484e4972b74a22ab66eaf7d779