processor: Parse legacy json DebugAnnotations during import
In crbug.com/1048815, we discovered that some begin/end events emitted
by Chrome contain the *same* legacy_json argument and expect the
contents of the dicts to be merged. We can accomplish this by exploding
the json dicts into separate args rows in the storage during import.

When exporting, we will rebuild the dict based off the rows. This
behaves exactly as the import/export of NestedValue proto
DebugAnnotations.

Bug: chromium:1048815
Change-Id: Id19c47e540484ebe3f40693513f2ca60998c3516