TraceProcessor: fix trace_bounds computation
The computation code was not taking into account the
duration of the last slice. So in a trace like this:

[   slice1   ][    slice 2   ][  slice 3  ]
|                             |
   Bounds before this CL.

This problem showed up only when analyzing ninja
logs, where the last event went missing.
Ironically this problem never shows up in most
ftrace-based traces, because after every slice
there is almost always a sched_wakeup event that
extends the bounds to the right values.

Bug: 149483774
Test: TraceProcessorIntegrationTest.TraceBounds
Change-Id: I202d1c9ce105f20e40a4eb3424e76f2a773615fa