TraceProcessor: add ninja log parsing
Adds a parser for ninja build logs. This is the
moral equivalent of https://github.com/nico/ninjatracing.
This works only as long as ninja doesn't hit recompaction.
The log seems to stop making sense after re-compaction because
it becomes impossible to reconstruct the boundaries of
each build invocation.
I eye-balled this vs ninjatracing on the same trace and
the results seem to make sense.

Bug: 149483774
Test: TraceProcessorIntegrationTest.NinjaLog
Change-Id: Ice9ac9d9fee6a94d78fb0fec997b3b3d36d3c02b