shmem: Require memfd in in-tree android builds
All kernel versions used on devices supported by in-tree builds should
support memfd, so require it when allocating shared buffers and check
for seals when receiving a buffer.

In out of tree builds, we continue to check for kernel support first,
so we can sideload perfetto onto devices with older kernels.

Bug: 148841422
Change-Id: I71412d8881c3c0c4312c8b40b830ceb3b75afa6e