traced_perf: enable the static libunwindstack cache
Arbitrary cache_enabled_ toggling seems not safe in a multithreaded
environment, but would still work for a single-threaded unwinder (just
that the toggling would have to be moved to be owned by the unwinder
thread).

The DLOG addendum is to make sure that the FormatFrame isn't evaluated
(I've seen it show up on release build profiles).

Bug: 144281346
Change-Id: I4af61cbb236e148341c09c30164e5664fb935752