tp: fix parsing of counter events from Chromium
Chromium counter events on Android can have an additional "category
group" marker appeneded to the end of the systrace counter event. This
currently causes parsing failures as we don't expect anything after the
value.

Fix the parsing to actually parse the group (even though we don't
actually plan on doing anything with it).

Bug: 149804989
Change-Id: I59bd709cf1bdf34c21bb6edb65118cebfbf20db9