Fix FtraceProcfsIntegrationTest.EnableDisableEvent
This test attempted to read back the trace while tracing
continued and relied being able to 'catch up' with the writer to
terminate. This did not always happen resulting in timeouts.

Modify the test to disable tracing before attempting to read back
the trace. Thanks to rsavitski@ for helping debug this.

Change-Id: I0a7610f01d1de62be21d6299d944b461fa795caf