traced_perf: unwind queue type
In the perf profiler, there's a need for samples to be enqueued between
the reader, and the unwinder. There will be a single writer thread, and
a (separate) single reader thread.

The queue is implemented as a basic ring buffer of an arbitrary type
(std::array<T, N> + atomic rd/wr pointers). The implementation is
generic, but I've kept the type with a dedicated name (UnwindQueue)
until we find a need to reuse it.

Likewise, the read and write primitives are skewed towards the expected
use-case: writer commits entries one by one, reader can consume a batch
of contiguous entries.

See follow-up patch for more context.

Bug: 144281346
Change-Id: I83d7d6100f5cf52fa91c93d6c75624af6b788b7e