traced_perf: allow slow proc-fd lookups to recover from expired state
Previously, slow proc-fd lookups would be marked as expired, and all
outstanding/future samples for that pid would be discarded. This was
written under the assumption that all requests that can succeed, will
succeed within the timeout. This patch allows for slow lookups to
recover from the expired state. So even if we have discarded some
samples in the mean time, future ones can be unwound.

This should be a slight improvement for long tracing sessions, where
losing some leading entries is not a big deal, compared with losing all
of them. As well as for CTS tests on sloooooow (read: cuttlefishy)
platforms.

This does however change how final the process state verdicts are for a
given tracing session. One could argue that for a given pid,
consistently skipping either none or all samples is easier to reason
about.

There are also some new edge cases with the combination of pid reuse and
concurrent data sources (since a data source can steal another's
proc-fds if it's also eligible to adopt them, which now includes expired
state). But I believe the edge cases should be both rare, and nothing
will go horribly wrong even if we hit them.

What do you think, worth it? I think it's still worth recovering for the
sake of longer traces, even if the initial request gets delayed by a
janky system state.

Bug: 144281346
Change-Id: I7176161efd5cb5c643d590a8ccc1a347ae7fde53