Use base::Subprocess in tests and tools
This CL changes the behavior of the cmdline and trigger
e2e tests. Instead of running the "perfetto" and
"trigger_perfetto" commands in a forked environment
(without exec()), exec the real binary.
fork() turns out to be super-fragile because forked
processes inherit the main process pipes that are
used to synchronize the test sequencing. By keeping
one end open, they prevent the main test process to
see the pipe EOF.

Change-Id: Ibc7838ee44fa3a2fa7c4d9fb24066de543a77e19