trace_processor: rewrite raw table to systrace conversion code
This CL overhauls the code which converts the lines of the raw table
into legacy systrace format.

Until this point, we were using a very fragile mechanism where we relied
on the fact that the ordering of proto fields == the ordering in the
args table. However, after aosp/1232520 we started sorting args based on
key id. While generally this won't cause breakage, in some subtle cases
(e.g. for binder_transaction), this was causing a reordering of the
fields and thus crashes on debug and wrong output on release.

Change this system to properly create a mapping between the order in
the proto and the args table. We assume that for each event type, we
will always get the same mapping of rows - this is a reasonable
assumption as we do not expect the fields of ftrace to change
mid-trace. Caching the ordering is necessary to get good performance.

While there is a small performance hit by doing this, it is pretty
negligible in the grand scheme of things (1% over for a 16GB trace).

Bug: 150919723
Change-Id: I5b8eff0c2f528f882428019bd9f448fed738733c