Acknowledge flushes / stop with unknown data sources.
This happens for fork-mode heapprofd, when a data source config is
rejected because of extra guardrails. We still need to acknowledge the
flush to prevent traced from waiting for the timeout.

Test: atest HeapprofdCtsTest on blueline-user

This is a cherry-pick of 3529fdda91842af29a7efd863fded778ce96975f.

Bug: 152368233
Change-Id: I5db7bc89f1d6904fe809795feb1eb206c2093fdf
Merged-In: I5db7bc89f1d6904fe809795feb1eb206c2093fdf