tp: improve diff testing script when tp returns error retcode
This CL changes the diff testing script to fail the test if trace
processor returned a non-zero error code. Before now, this would cause
silent failures if the output and expected file were both unlikely (this
happens when the test is first being written) and led to confusing
passes of tests which should have failed.

Change-Id: I90bfc4de746355b511acd8709572aaf68c6b458a