tp: fix returning utid of freed thread in rss stat event
Once a process is freed, it's mm struct (and thus id) can be reused.
We assumed that we would see an rss_stat event for the process reusing
the struct before other processes modified it but this doesn't seem to
be the case in practice. This causes negative durations in the memory
metric as we make use of the end ts of the process as the termination of
memory events.

Fix this by double checking the utid in the map is alive before
returning it.

Bug: 152521411
Change-Id: I92d091db33314b9b797b6d867e4c8e4eaf0a20ab