perfetto_cmd: if unique session is given, set comm to "p-session_name"
Why not argv0: (a) more involved, as it requires writing directly to
some process memory, (b) it's harder for tools to notice changes in
process names than in comm, (c) process name of "perfetto" is still
generally valid, few people will need to dig deeper (and then the thread
name provides extra info).

The core function is intentionally defined only on platforms that it is
implemented on.

I guessed at the OSX branch given the TTR example.

Change-Id: I2aaeb0edd2171ea3dfd3bd8acb7cbc087e79f5e3