service: allow to specify a file path with write_into_file
Until now write_into_file was based on the consumer passing a file
descriptor over the IPC socket. This doesn't work for Android
Graphics Debugger, which runs on the host and uses the IPC protocol
over adb forward (which doesn't support passing FDs).
Add an option to pass a path instead. This will require a matching
selinux change as the service is not allowed to create files at the
moment.
The feature conservatively rejects over-writing existing files, to
prevent being misused. This feature is only exposed to the consumer
socket (i.e. shell) which is trusted.

Bug: 153519149
Change-Id: I03d793a2630018e60b4724e16bca5f402fdc225b