Tracing protocol: chunk QueryServiceStateResponse
The QueryServiceStateResponse IPC had a design oversight:
it was assuming that the response will fit into one IPC
message. This is not true when a lot of data sources with
large descriptors are registered. Each DataSourceDescritor
can be up to the size of an IPC message each.
Fix by chunking the response, splitting at the data source
level.
This change is binary compatible with the previous version.
The service still ensures that each chunk is a valid (but
incomplete) TracingServiceState response, so old clients
will still see the first response (which in most cases fits).

Bug: 153142114
Change-Id: I8256a440ddea1973400d8b2dd1d0496358bd2cd7