traced_perf: support incremental state clears
For resetting incremental sequence state in ring buffer traces. The only
quirky thing I'm doing is not clearing the output trackers of concurrent
data sources when dropping the global trie. Technically we know that
that data should not be used anymore, but I don't like the cross-talk of
one data source mutating the others, and the "leak" should be trivially
low.

Bug: 144281346
Change-Id: Id366f9b87a41a6452c2336634166960662c107f3