IPC: Add Sync() method to synchronize Producer and Service
Use this to fix test flakiness in the e2e test.
The problem is the following: most of the tests that use
FakeProducer assume that the producer has registered its
FakeDataSource before StartTracing is called. That's not
guaranteed to happen.
This CL makes use of the Sync() method to enforce that.

Test: perfetto_integrationtests --gtest_repeat=1000
Change-Id: I6b5d384b624229fd795e6d434c77516136b6fa4e