tp: move parsing of systeminfo from parse to tokenization
SystemInfo does not have a timestamp as it's synthesized by traced.
Instead, it should be treated as a packet which appears before all data
packets.

Treat the systeminfo like any other context packet and parse it during
tokenization time so that we have to context inside available when
parsing real trace packets.

Bug: 155308466
Change-Id: Iaf767a2d93819531cf51e83ab0ab4f72a2b3d85b