ui: Fix nesting of slice argument table
This patch makes the arguments of a slice be displayed directly in the
"Slice Details" table rather than as standalone <tr> <td> tags (with no
containing <table>). We also remove the "Args" heading since looks a bit
odd being an extra row in the table rather than a real heading (which
would take more space).

Change-Id: I2217e8b64b38a368478d1acedb10baf6224ec7e0