heapprofd_client: systematically leak state in an atfork handler
Proposed mitigation for dealing with processes that fork while
some thread is inside the hooks (and thus potentially holding
locks and/or otherwise mutating Client state).

As discussed, there's no good way of dealing with all possiblities
(perhaps short of re-structuring the client hooks altogether), but this
version is at the very least minimally invasive to the host process.

See the code comments for more details/rationale. Though I'm omitting
many details as any proper description of fork/clone issues would
require writing a book.

Tested manually with a mini binary that reproduces a fork during a
thread being inside malloc, but didn't do much soak testing for
"heap profiling still works in general".

This might not play well with sanitizers, but we'll have to see.

Bug: 155193946
Change-Id: I8e35e204beee92023811c8b6552ff2114f79a11a