ui: Introduce analyze page with multiline SQL input box
- The input box on analyze page is decoupled from the quick SQL input box
on the topbar.
- It resizes automatically as user enters more lines up until roughly
10 lines, then shows scroll bar.
- It handles tabs as 2 spaces instead of losing focus.
- If the user resizes the textbox by dragging the resize handle, it
respects that resize value instead of auto resizing.

Unfortunately clicking on a row in the analyze page does not take you
to the slice in the timeline view, as it does in from the output of the
SQL query created in topbar. I couldn't find a clean way of making it
work, and will address it in a later CL.

Bug:149206686
Change-Id: If0c04abcbfcca53b3619f850f03557e44e3a747a