clock: Support specifying an authoritative trace clock in the config.
Chrome traces use CLOCK_MONOTONIC as their trace clock. If the trace
only contains Chrome data, it doesn't make sense to convert its
timestamps to CLOCK_BOOTTIME.

To enable instructing trace processor not to attempt this conversion,
this patch adds a way to specify the clock domain that should be used
for the trace as part of the trace config. It's written into the trace
file's header (as part of a ClockSnapshot) by the service.

This also requires moving the BuiltinClock(s) enum into
protos/perfetto/common.

Bug: chromium:1060400, chromium:1085002

Change-Id: I1c152b37905e161cb2756e15f687978dc5e14f58