site: improve layout of home page on mobile devices
Improve the sizing of <h1> <h2> and <img> of the above-the-fold
section of the home page.
I manually tested the new layout on the various form factors.
Also update the links in the stack svg picture and run optipng
on the home picture, reducing its size by 20%.

Preview in https://staging-dot-perfetto-site.uc.r.appspot.com/

Change-Id: I53c3f6b743c3d3858da84c54fd695e7008854153