tp: add metatracing
This CL adds metatracing support to trace processor using the existing
metatracing API in base. This includes plumbing to shell with the -m
flag allowing metatracing of shell.

We use a new implementation of metatracing as the one in base does not
have quite the right shape for use in trace processor (e.g. the use of
task runners, the lack of args support etc)

Change-Id: Ib052d8d2d623f18e95c8614448975617d914dcca