service: Capture clock snapshots periodically while tracing
Refactors periodic snapshots in the service in such a way that we can
keep track of clock drift that happens while tracing even if we are not
streaming the trace to a file/consumer, which is particularly important
in ring-buffer mode (see bug).

We post recurring tasks to update the latest clock snapshot if new drift
has been detected and emit the latest clock snapshot on the next call
to ReadBuffers().

Also makes the snapshotting interval configurable.

Bug: 157467831
Change-Id: I3c8bc4e0c76f806fd248c36de6749fd084998c83