TraceProcessor: support resolution of non-strictly monotonic clocks
The clock sync logic was treating non-stricly monotonic clocks as
non-monotonic, refusing to use them on the clock resolution path.
This is incorrect and actually can happen with CLOCK_MONOTONIC_COARSE,
which is clearly monotonic (as the name suggests) but not stricly.
What's failing is the case of two snapshots taken back-to-back, where
BOOTTIME and the other clocks advance, but not MONOTONIC_COARSE.

Bug: 158182858
Test: ClockTrackerTest.NonStrictlyMonotonic
Change-Id: I412f4143b8853b7086c2c8c53473488a640af779