Fix query to exclude tracks with no time_in_state
https://r.android.com/1331924 excluded tracks when a thread has no
time_in_state on any of the core types. Due to ambiguity in the WHERE
clause, it kept tracks for the big core even if only the little one
recorded any time_in_state. This CL fixes this to exlude on the track
level.

This CL also makes the test more realistic by simulating core types.

Bug: 157216280
Test: tools/diff_test_trace_processor.py
Change-Id: I41d78ca9314a8fe44201c144c540d9a6afe59d1a