tp: allow span join to propogate constraints on timestamps
This CL does two things:
1. For ts <= x constraints on span joined tables, it propogates these
constraints directly to the source tables. This is valid because any
slices in the span joined table which occur before x necessarily have to
orginate from events which also occur before x.
2. We add a new constraint function source_geq which forces span join to
propogate a ts >= x constraint to the two source tables.

The usage of this is like so:
select ts, dur
from span
where source_geq(ts, x)
This causes a ts >= x constraint to be issued to the two tables which
are being span joined.

The intention behind both of these changes is to remove usage of the
window table from the Perfetto UI - these changes are necessary for this
to happen.

Change-Id: I68ef7eefe682fbb939920f69f13a312c73182355