ui: bring new quantization to regular counter tracks
This CL introduces the new approach to quantization [1] to ordinary
counter tracks.

As part of this, we also remove the gpu frequency track implementation
which is no longer necessary as, now everything is keyed by track id,
the regular counter track implementation has the correct shape to also
show gpu frequency events.

Change-Id: I9da489ae9dbe2513e2b57bbef6080f19f715917e