GN: support cross-compilation with same cpu and os
Before this CL we use two distinct toolchains only when
target_os != host_os or target_cpu != host_cpu.
This isn't enough for CrOS where both the host and target are x64
(x86_64 really), but they need different toolchain arguments.
This CL further conditions the is_cross_comping logic looking at
target_triplet.
In order to do so, the default initialization of target_triplet
is moved to BUILDCONFIG.gn (which is the one in charge of
setting is_cross_compiling) and the existing logic is moved
to _target_triplet (with underscore).

Bug: 147789115
Change-Id: I5c239e9cb35c0c5779ef8c615eaa9de309637544