perf: bump max sampled stack, + avoid reparse when stack is truncated
Certain tasks (e.g. traced_probes) often have deep stacks (primarily due
to on-stack data), while perf record limit (64k) limits the amount of
stack that the kernel can save in a sample.

If we fail to fully unwind a sample with a "saturated" stack sample,
skip reparsing maps and attempting a reunwind, since the cause of the
failed unwind is likely the truncation.

Still deliberating whether the saturation is worth recording in the
trace - it helps to disambiguate the likely cause of the partial unwind,
but there's no existing field where it'd fit.

Change-Id: I3c67b7b2a6546dcfceea94c8cc63164d8defe587