perf: properly reset regs on retry, + do not reparse after 2nd attempt
Primary objective is to fix the oversight where the registers weren't
being reset properly during unwinding retries, causing us to continue at
the broken frame (as the pc has been advanced).

Also, tone down on the maps reparses by not attempting them after the
2nd attempt, even if it reached an invalid mapping.

While here, reorganized the unwinding logic to avoid the loop. Now
constructing a fresh Unwinder object (should be ~free) for every
attempt, arguing that this is simpler to reason about. The lambda's
return type helps avoid having to reason about the lifetime of the
Unwinder object vs the state it references (for example, it remembers
the raw ptr to the regs object, so returning the Unwinder from the
lambda would be a bug).

Bug: 160697674
Change-Id: I461fa63a541686a276d13becb17cc501c6f4a7b4