tp: add support for opting out of query output checks in shell
This CL adds support for telling shell to ignore the output of a SELECT
RUN_METRIC query without using the horrible CREATE TABLE hack which
hides syntax errors. We do this by naming the single null returned from
RUN_METRIC as "suppress_query_output" which is a hint to shell to
not count this query towards multiple output detection.

Change-Id: I7677e6421459272e9a359a3bc9d7b7f134815776