tp: merge test/metrics into test/trace_processor
These were separate becaue they developed somewhat independently but
have grown increasingly the same over time. Because we have a consistent
naming system for query tests (file ends with .sql), just merge the two
folders together and rely on the extension to define whether the test is
a query test or metric test.

Change-Id: Ia73960739260971aa40352970d765ef27eb00271